Move and rename some important classes
