+++ src/main/java/seedu/todo/ui/view/CommandErrorView.java
@@ -11,7 +11,7 @@ import seedu.todo.commons.core.LogsCenter;
 import seedu.todo.ui.util.FxViewUtil;
 import seedu.todo.model.ErrorBag;
 import seedu.todo.ui.UiPart;
-import seedu.todo.ui.UiPartLoader;
+import seedu.todo.ui.util.UiPartLoaderUtil;
 import seedu.todo.ui.util.ViewGeneratorUtil;
 import seedu.todo.ui.util.ViewStyleUtil;
 
@@ -46,7 +46,7 @@ public class CommandErrorView extends UiPart {
      * @return an instance of this class
      */
     public static CommandErrorView load(Stage primaryStage, AnchorPane placeHolder) {
-        CommandErrorView errorView = UiPartLoader.loadUiPart(primaryStage, placeHolder, new CommandErrorView());
+        CommandErrorView errorView = UiPartLoaderUtil.loadUiPart(primaryStage, placeHolder, new CommandErrorView());
         errorView.addToPlaceholder();
         errorView.configureLayout();
         errorView.hideCommandErrorView();
